requirements.txt
boto3==1.34.148
botocore==1.34.148
kafka_python_ng==2.2.2
pandas==2.2.2
pendulum==3.0.0
schedule==1.2.2
SQLAlchemy==1.4.53
